refactor(SubscrptionList): extract useSubscriptions hook
