+++ components/SubscriptionList.tsx
@@ -1,56 +1,15 @@
-import { AppContext } from "pages/_app";
-import { useContext, useEffect, useState } from "react";
-
-interface Subscription {
-  id: number;
-  feedName: string;
-}
-
-const ERROR_MESSAGE =
-  "Oops! There was an error loading your subscriptions. Please try again later.";
+import { useSubscriptions } from "hooks/useSubscriptions";
 
 export const SubscriptionList = () => {
-  const { user } = useContext(AppContext);
-  const [subscriptions, setSubscriptions] = useState<Subscription[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [errorMessage, setErrorMessage] = useState("");
-
-  useEffect(() => {
-    setErrorMessage("");
-
-    if (!user?.token) {
-      return;
-    }
-
-    setIsLoading(true);
-
-    fetch("/api/subscriptions", {
-      method: "POST",
-      body: JSON.stringify({ authToken: user?.token }),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        if (data.status !== 200) {
-          setErrorMessage(ERROR_MESSAGE);
-          return;
-        }
-
-        setSubscriptions(data.subscriptions);
-      })
-      .catch(() => {
-        setErrorMessage(ERROR_MESSAGE);
-      })
-      .finally(() => {
-        setIsLoading(false);
-      });
-  }, [user?.token]);
+  const { subscriptions, subscriptionsLoading, subscriptionError } =
+    useSubscriptions();
 
   return (
     <section className="manage-feeds-section">
       <h2>Your subscriptions</h2>
 
-      {errorMessage && <p className="error">{errorMessage}</p>}
-      {isLoading && <p>Loading...</p>}
+      {subscriptionError && <p className="error">{subscriptionError}</p>}
+      {subscriptionsLoading && <p>Loading...</p>}
 
       {subscriptions.length > 0 && (
         <ul>
